fixed 16-vs-32-bit issues that stopped some native compiles

+++ src/lex.c
@@ -88,7 +88,7 @@ static void out_record_read(unsigned record)
 /* Report the current record/offset */
 unsigned long out_tell(void)
 {
-	return (outrecord << 8) | outlen;
+	return ((unsigned long)outrecord << 8) | outlen;
 }
 
 /* Go to a given record/offset from before */
